@@ -35,13 +35,13 @@ extension IntExt on num? {
   String toLikes() {
     var num = this ?? 0;
     if (num > 999 && num < 99999) {
-      return "${(num / 1000).toStringAsFixed(1)}K";
+      return '${(num / 1000).toStringAsFixed(1)}K';
     } else if (num > 99999 && num < 999999) {
-      return "${(num / 1000).toStringAsFixed(0)}K";
+      return '${(num / 1000).toStringAsFixed(0)}K';
     } else if (num > 999999 && num < 999999999) {
-      return "${(num / 1000000).toStringAsFixed(1)}M";
+      return '${(num / 1000000).toStringAsFixed(1)}M';
     } else if (num > 999999999) {
-      return "${(num / 1000000000).toStringAsFixed(1)}B";
+      return '${(num / 1000000000).toStringAsFixed(1)}B';
     } else {
       return num.toString();
     }
